docker/discord_bot/requirements.txt
discord.py>=2.0.0
boto3>=1.26.0
python-dotenv>=0.21.0
requests>=2.28.0